Lesson 3 - Intents
The project implements a few examples of how to use implicit and explicit intents. For the latter, we have used ListView in combination with ListAdapter to provide a list of courses. The user can select a specific course and pass it as a parameter to the following activity to visualize further details.